[04] Add ability to change year

+++ 04 - Sparklines/src/js/main.js
@@ -1,28 +1,26 @@
-function loadData(){
+function loadData() {
   //window.localStorage.clear();
-  let sparkline = getVal('sparkline');
-  if(sparkline.length==0)
-  {
-    let box = {
-      id: 'pair00',
-      from: 'BTC',
-      to: 'ETH',
-      year: '2021'
+  let sparkline = getVal("sparkline");
+  if (sparkline.length == 0) {
+    const box = {
+      id: "pair00",
+      from: "BTC",
+      to: "ETH",
+      year: "2021",
     };
     sparkline.unshift(box);
-    displayBox(box.id,box.from,box.to,box.year);
+    displayBox(box.id, box.from, box.to, box.year);
     window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
-  }
-  else sparkline.forEach(box=>{
-    displayBox(box.id,box.from,box.to,box.year);
-  });
+  } else
+    sparkline.forEach((box) => {
+      displayBox(box.id, box.from, box.to, box.year);
+    });
 }
 
-function getVal(item){
+function getVal(item) {
   let val = window.localStorage.getItem(item);
-  if(val == null)
-    return [];
-    return JSON.parse(val);
+  if (val == null) return [];
+  return JSON.parse(val);
 }
 
 function warning(msg) {
@@ -32,63 +30,79 @@ function warning(msg) {
   }, 900);
 }
 
-function invert(node){
-  let id = node.parentNode.id;
-  var from,to,year;
-  let sparkline = getVal('sparkline');
-  sparkline.forEach((box)=>{
-    if(box.id == id)
-    {
+function invert(node) {
+  const id = node.parentNode.id;
+  let from, to, year;
+  let sparkline = getVal("sparkline");
+  sparkline.forEach((box) => {
+    if (box.id == id) {
       let temp = box.from;
       box.from = box.to;
       box.to = temp;
-      from = `${box.from}`;
-      to = `${box.to}`;
+      from = box.from;
+      to = box.to;
       year = box.year;
       return;
     }
-  })
+  });
   window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
   let card = document.getElementById(id);
-  card.querySelector('#pair').innerHTML = `${from}/${to}`;
-  getData(from,to,year,id);
+  card.querySelector("#pair").innerHTML = `${from}/${to}`;
+  getData(from, to, year, id);
 }
 
-function getData(from,to,year,id) {
+function change(node){
+  const id = node.parentNode.id;
+  let from, to, year;
+  year = node.value;
+  let sparkline = getVal("sparkline");
+  sparkline.forEach((box) => {
+    if (box.id == id) {
+      from = box.from;
+      to = box.to;
+      box.year = year;
+      return;
+    }
+  });
+  window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
+  let card = document.getElementById(id);
+  getData(from, to, year, id);
+}
+
+function getData(from, to, year, id) {
   //console.log("first fn");
   let prices = [];
   fetch(`data/${from}/${year}.json`)
     .then((response) => response.json())
-    .then((data) => 
-    {
-      prices = data['prices'];
+    .then((data) => {
+      prices = data["prices"];
       return fetch(`data/${to}/${year}.json`)
-      .then((response) => response.json())
-      .then((data) => getValues(prices,data['prices'],id)); 
-    })
+        .then((response) => response.json())
+        .then((data) => getValues(prices, data["prices"], id));
+    });
 }
 
-function getValues(datafrom,datato,id) {
+function getValues(datafrom, datato, id) {
   //console.log('hihhii',datafrom,datato,id);
-  var data = [];
-  datafrom.forEach((Rate,i) =>{
-    let price = datafrom[i]/datato[i];
+  let data = [];
+  datafrom.forEach((Rate, i) => {
+    let price = datafrom[i] / datato[i];
     if (price < 1)
-    price = price.toFixed(1 - Math.floor(Math.log(price) / Math.log(10)));
+      price = price.toFixed(1 - Math.floor(Math.log(price) / Math.log(10)));
     else price = price.toFixed(2);
     data.push(parseFloat(price));
   });
-  let parent = document.getElementById(id);
-  let first = (data[0]),
-    last = (data[data.length - 1]);
-  let min = (data[0]),
-    max = (data[0]);
+  const parent = document.getElementById(id);
+  let first = data[0],
+    last = data[data.length - 1];
+  let min = data[0],
+    max = data[0];
   data.forEach((Rate, i) => {
     if (Rate < min) min = Rate;
     if (Rate > max) max = Rate;
   });
-  parent.querySelector('#low').innerHTML = `${min}`;
-  parent.querySelector('#high').innerHTML = `${max}`;
+  parent.querySelector("#low").innerHTML = `${min}`;
+  parent.querySelector("#high").innerHTML = `${max}`;
   let change = ((last - first) / first) * 100;
   if (Math.abs(change) < 1)
     change = change.toFixed(1 - Math.floor(Math.log(change) / Math.log(10)));
@@ -105,56 +119,60 @@ function getValues(datafrom,datato,id) {
     parent.querySelector("#change").style.color = "green";
   }
 
-  //Sparkline 
+  //Sparkline
   //fillColor: '' lineColor: ''
-    
-  $(`#${id} .dynamicsparkline`).sparkline(data, {height: 150,width: 250, lineWidth: 1,chartRangeMin: 0});
+
+  $(`#${id} .dynamicsparkline`).sparkline(data, {
+    height: 150,
+    width: 250,
+    lineWidth: 1,
+    chartRangeMin: 0,
+  });
 }
 
-function remove(node){
-  let id = node.parentNode.id;
-  let sparkline = getVal('sparkline');
-  sparkline.forEach((box,i)=>{
-    if(box.id == id)
-    {
-      sparkline.splice(i,1);
-      return;
-    }
-  })
-  window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
-  node.parentNode.remove();
+function remove(node) {
+  if (window.confirm("Are you sure?")) {
+    let id = node.parentNode.id;
+    let sparkline = getVal("sparkline");
+    sparkline.forEach((box, i) => {
+      if (box.id == id) {
+        sparkline.splice(i, 1);
+        return;
+      }
+    });
+    window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
+    node.parentNode.remove();
+  }
 }
 
-function displayAdd(){
-  document.getElementById('addform').style.display = 'flex';
+function displayAdd() {
+  document.getElementById("addform").style.display = "flex";
 }
 
-function add(node){
-  let from = node.querySelector('#c1').value;
-  let to = node.querySelector('#c2').value;
-  let year = node.querySelector('#period').value;
-  let count = window.localStorage.getItem('count');
-  if(count == null)
-  count = 0;
-  let sparkline = getVal('sparkline');
+function add(node) {
+  let from = node.querySelector("#c1").value;
+  let to = node.querySelector("#c2").value;
+  let year = node.querySelector("#period").value;
+  let count = window.localStorage.getItem("count");
+  if (count == null) count = 0;
+  let sparkline = getVal("sparkline");
   let box = {
     id: `pair${count}`,
     from: from,
     to: to,
-    year: year
+    year: year,
   };
   count++;
   sparkline.push(box);
   //console.log(sparkline);
   window.localStorage.setItem("sparkline", JSON.stringify(sparkline));
   window.localStorage.setItem("count", count);
-  document.getElementById('addform').style.display = 'none';
-  displayBox(box.id,from,to,year);
+  document.getElementById("addform").style.display = "none";
+  displayBox(box.id, from, to, year);
 }
 
-function displayBox(id,from,to,year){
-  let boxHTML = 
-  `<div class="box" id="${id}">
+function displayBox(id, from, to, year) {
+  let boxHTML = `<div class="box" id="${id}">
   <span name="pair" id="pair">${from}/${to}</span>
   <img class="delete" src="img/delete.svg" alt="delete" onclick="remove(this)"></br>
   <img class="invert" src="img/swap.svg" alt="invert" onclick="invert(this)">
@@ -165,9 +183,14 @@ function displayBox(id,from,to,year){
       <div class="high"><img class="lhicon" src="img/high.svg" alt=""><span id="high">0</span></div>
   </div>
   <span class="dynamicsparkline"></span>
-  <span name="period" id="period">${year}</span>
+  <select name="period" id="period" onchange="change(this)">
+  <option value="2019">2019</option>
+  <option value="2020">2020</option>
+  <option value="2021">2021</option>
+  </select>
 </div>`;
-  let container = document.getElementById('container');
-  container.insertAdjacentHTML('afterbegin',boxHTML);
-  getData(from,to,year,id);
+  let container = document.getElementById("container");
+  container.insertAdjacentHTML("afterbegin", boxHTML);
+  document.getElementById(id).querySelector('#period').value = year;
+  getData(from, to, year, id);
 }
